+++ core/src/main/java/tech/pegasys/web3signer/core/service/jsonrpc/EthSendTransactionJsonParameters.java
@@ -35,6 +35,9 @@ public class EthSendTransactionJsonParameters {
   private String input;
   private BigInteger maxFeePerGas;
   private BigInteger maxPriorityFeePerGas;
+  private BigInteger maxFeePerBlobGas;
+  private String[] blobVersionedHashes;
+  private String[] blobs;
 
   @JsonCreator
   public EthSendTransactionJsonParameters(@JsonProperty("from") final String sender) {
@@ -87,6 +90,21 @@ public class EthSendTransactionJsonParameters {
     this.maxFeePerGas = decodeBigInteger(maxFeePerGas);
   }
 
+  @JsonSetter("maxFeePerBlobGas")
+  public void maxFeePerBlobGas(final String maxFeePerBlobGas) {
+    this.maxFeePerBlobGas = decodeBigInteger(maxFeePerBlobGas);
+  }
+
+  @JsonSetter("blobVersionedHashes")
+  public void blobVersionedHashes(final String[] blobVersionedHashes) {
+    this.blobVersionedHashes = blobVersionedHashes;
+  }
+
+  @JsonSetter("blobs")
+  public void blobs(final String[] blobs) {
+    this.blobs = blobs;
+  }
+
   public Optional<String> data() {
     if (data != null) {
       return Optional.of(data);
@@ -126,5 +144,17 @@ public class EthSendTransactionJsonParameters {
 
   public Optional<BigInteger> maxFeePerGas() {
     return Optional.ofNullable(maxFeePerGas);
+  }
+
+  public Optional<BigInteger> maxFeePerBlobGas() {
+    return Optional.ofNullable(maxFeePerBlobGas);
+  }
+
+  public Optional<String[]> blobVersionedHashes() {
+    return Optional.ofNullable(blobVersionedHashes);
+  }
+
+  public Optional<String[]> blobs() {
+    return Optional.ofNullable(blobs);
   }
 }